Just split impact module, since it's growing so much
